[refactor] Move reusable functions into module
